Reset input capacity when setting external shader program
Before this CL, externalShaderProgramInputCapacity was not reset when
the external shader program was reset (which occurs when  the
InputSwitcher switches to an input with a different ColorInfo). This is
due to a regression introduced in https://github.com/androidx/media/commit/64b08e8ecc5f180fb9f19e8506492291f3794f38.

PiperOrigin-RevId: 637869215