Add a test for `MediaCodecRenderer` handling of `IllegalStateException`
This is a regression test for the bug introduced in https://github.com/androidx/media/commit/bf55d3ca6a846f399b17d0859033fc3b900380da
which was manually spotted and fixed in https://github.com/androidx/media/commit/fe4dcb4ca1a27b658489550fbaeea49842a2d0c4.

Reverting the fix causes this test to fail.

This test is a bit hacky because we have to munge the stack trace of
the `IllegalStateException` to make it look like it was thrown from
inside `MediaCodec`. We deliberately do this 'badly' (e.g. using
`fakeMethod`) to avoid a future reader being confused by a
fake-but-plausible stack trace.

PiperOrigin-RevId: 652820878